requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
docopt==0.6.2
get==2019.4.13
idna==2.9
mysql-connector==2.2.9
pipreqs==0.4.10
post==2019.4.13
public==2019.4.13
query-string==2019.4.13
requests==2.23.0
urllib3==1.25.9
yarg==0.1.9
